Move update_exc to global language data utils
