+++ spacy/es/__init__.py
@@ -6,8 +6,8 @@ from os import path
 from ..language import Language
 from . import language_data
 from ..attrs import LANG
-from ..util import update_exc
 
+from ..language_data import update_exc
 from ..language_data import EMOTICONS
 from .language_data import ORTH_ONLY
 from .language_data import strings_to_exc
